Add SemiSL Mixin
