+++ src/otx/algorithms/classification/adapters/mmcls/configurer.py
@@ -3,12 +3,11 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
-import importlib
 from typing import Optional
 
 import torch
 from mmcv import build_from_cfg
-from mmcv.utils import Config, ConfigDict
+from mmcv.utils import ConfigDict
 
 from otx.algorithms import TRANSFORMER_BACKBONES
 from otx.algorithms.classification.adapters.mmcls.utils import (
@@ -16,10 +15,7 @@ from otx.algorithms.classification.adapters.mmcls.utils import (
     patch_evaluation,
 )
 from otx.algorithms.common.adapters.mmcv.configurer import BaseConfigurer
-from otx.algorithms.common.adapters.mmcv.utils import (
-    build_dataloader,
-    build_dataset,
-)
+from otx.algorithms.common.adapters.mmcv.semisl_mixin import SemiSLConfigurerMixin
 from otx.algorithms.common.adapters.mmcv.utils.config_utils import (
     InputSizeManager,
     get_configured_input_size,
@@ -285,48 +281,5 @@ class IncrClassificationConfigurer(ClassificationConfigurer):
             update_or_add_custom_hook(cfg, ib_loss_hook)
 
 
-class SemiSLClassificationConfigurer(ClassificationConfigurer):
+class SemiSLClassificationConfigurer(SemiSLConfigurerMixin, ClassificationConfigurer):
     """Patch config to support semi supervised learning for classification."""
-
-    def configure_hook(self, cfg):
-        """Update cfg.custom_hooks."""
-        super().configure_hook(cfg)
-        # Set unlabeled data hook
-        if self.training:
-            if cfg.data.get("unlabeled", False) and cfg.data.unlabeled.get("otx_dataset", False):
-                self.configure_unlabeled_dataloader(cfg)
-
-    @staticmethod
-    def configure_unlabeled_dataloader(cfg: Config):
-        """Patch for unlabled dataloader."""
-
-        model_task = {"classification": "mmcls", "detection": "mmdet", "segmentation": "mmseg"}
-        if "unlabeled" in cfg.data:
-            task_lib_module = importlib.import_module(f"{model_task[cfg.model_task]}.datasets")
-            dataset_builder = getattr(task_lib_module, "build_dataset")
-            dataloader_builder = getattr(task_lib_module, "build_dataloader")
-
-            dataset = build_dataset(cfg, "unlabeled", dataset_builder, consume=True)
-            unlabeled_dataloader = build_dataloader(
-                dataset,
-                cfg,
-                "unlabeled",
-                dataloader_builder,
-                distributed=cfg.distributed,
-                consume=True,
-            )
-
-            custom_hooks = cfg.get("custom_hooks", [])
-            updated = False
-            for custom_hook in custom_hooks:
-                if custom_hook["type"] == "ComposedDataLoadersHook":
-                    custom_hook["data_loaders"] = [*custom_hook["data_loaders"], unlabeled_dataloader]
-                    updated = True
-            if not updated:
-                custom_hooks.append(
-                    ConfigDict(
-                        type="ComposedDataLoadersHook",
-                        data_loaders=unlabeled_dataloader,
-                    )
-                )
-            cfg.custom_hooks = custom_hooks
