+++ src/otx/algorithms/segmentation/adapters/mmseg/configurer.py
@@ -3,20 +3,16 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
-import importlib
 import os
 from collections import OrderedDict
-from typing import Any, Dict, List, Optional
+from typing import Any, List, Optional
 
 import torch
 from mmcv.runner import CheckpointLoader
 from mmcv.utils import Config, ConfigDict
 
 from otx.algorithms.common.adapters.mmcv.configurer import BaseConfigurer
-from otx.algorithms.common.adapters.mmcv.utils import (
-    build_dataloader,
-    build_dataset,
-)
+from otx.algorithms.common.adapters.mmcv.semisl_mixin import SemiSLConfigurerMixin
 from otx.algorithms.common.adapters.mmcv.utils.config_utils import (
     InputSizeManager,
     get_configured_input_size,
@@ -229,16 +225,8 @@ class IncrSegmentationConfigurer(SegmentationConfigurer):
         update_or_add_custom_hook(cfg, task_adapt_hook)
 
 
-class SemiSLSegmentationConfigurer(SegmentationConfigurer):
+class SemiSLSegmentationConfigurer(SemiSLConfigurerMixin, SegmentationConfigurer):
     """Patch config to support semi supervised learning for semantic segmentation."""
-
-    def configure_hook(self, cfg):
-        """Update cfg.custom_hooks."""
-        super().configure_hook(cfg)
-        # Set unlabeled data hook
-        if self.training:
-            if cfg.data.get("unlabeled", False) and cfg.data.unlabeled.get("otx_dataset", False):
-                self.configure_unlabeled_dataloader(cfg)
 
     def configure_task(self, cfg: ConfigDict, **kwargs: Any) -> None:
         """Adjust settings for task adaptation."""
@@ -246,45 +234,3 @@ class SemiSLSegmentationConfigurer(SegmentationConfigurer):
 
         # Remove task adapt hook (set default torch random sampler)
         remove_custom_hook(cfg, "TaskAdaptHook")
-
-    @staticmethod
-    def configure_unlabeled_dataloader(cfg: ConfigDict) -> None:
-        """Patch for unlabled dataloader."""
-
-        model_task: Dict[str, str] = {
-            "classification": "mmcls",
-            "detection": "mmdet",
-            "segmentation": "mmseg",
-        }  # noqa
-        if "unlabeled" in cfg.data:
-            task_lib_module = importlib.import_module(f"{model_task[cfg.model_task]}.datasets")
-            dataset_builder = getattr(task_lib_module, "build_dataset")
-            dataloader_builder = getattr(task_lib_module, "build_dataloader")
-
-            dataset = build_dataset(cfg, "unlabeled", dataset_builder, consume=True)
-            unlabeled_dataloader = build_dataloader(
-                dataset,
-                cfg,
-                "unlabeled",
-                dataloader_builder,
-                distributed=cfg.distributed,
-                consume=True,
-            )
-
-            custom_hooks = cfg.get("custom_hooks", [])
-            updated = False
-            for custom_hook in custom_hooks:
-                if custom_hook["type"] == "ComposedDataLoadersHook":
-                    custom_hook["data_loaders"] = [
-                        *custom_hook["data_loaders"],
-                        unlabeled_dataloader,
-                    ]
-                    updated = True
-            if not updated:
-                custom_hooks.append(
-                    ConfigDict(
-                        type="ComposedDataLoadersHook",
-                        data_loaders=unlabeled_dataloader,
-                    )
-                )
-            cfg.custom_hooks = custom_hooks
